Moved Header and Sample to other package. Added option to ExponentialMovingAverage that let's you choose whether to include anomalous scores in calculation of average or not.
